+++ prisma/migrations/20220302191841_removed_currency_from_order/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Order" DROP COLUMN "currency";