+++ prisma/migrations/20220302200727_changed_currency_to_required_in_symbol_profile/migration.sql
@@ -1,0 +1,5 @@
+-- Set default value
+UPDATE "SymbolProfile" SET "currency" = 'USD' WHERE "currency" IS NULL;
+
+-- AlterTable
+ALTER TABLE "SymbolProfile" ALTER COLUMN "currency" SET NOT NULL;